refactor(runtime): standardize model storage under artifacts directory
